fix reload infinito

+++ src/components/Login/sair.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 
 export default function Sair() {
+
+  if (sessionStorage.length > 0) {
     // deletar o cookie também
     sessionStorage.clear();
-    
-    // reload infinito
     window.location.reload();
+  }
 
-    return(
-        <h1>Deslogado!</h1>
-    )
+  return (
+    <h1>Deslogado!</h1>
+  )
 }
